chore: update package.json with new version and dependencies
The new version is 0.1.2, and the updated peerDependencies now include prettier version ^3.3.3. The dependencies section has removed prettier as it was previously listed under both peerDependencies and dependencies.

Also note that "typescript" dependency remains at version "^5.0.0". There's a discrepancy in the commit message since the typescript version is still mentioned, whereas it should be updated to include the new prettier version correctly in the peers section.